requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
pydantic-settings==2.1.0
mysql-connector-python==8.3.0
PyPika==0.48.9
fastapi-restful==0.5.0
typing-inspect==0.9.0
passlib==1.7.4
pudb==2024.1
python-jose==3.3.0
cryptography==42.0.5
python-multipart==0.0.9
